Add screen bounds and viewpoint calc to Atlas
Atlas now tracks four coordinates for the L/R and T/B edges of the
screen, comoving with the current tile. Also calculates their common
ancestor to use as a viewpoint for drawing.

This requires a geometry for screen dimensions when initialising an
atlas. In future the atlas will need to be reupdated for each screen
size change.