Major change to wdraw_atlas concept
Drawing an atlas now uses its viewpoint Coordinate. The wdraw_chart_with
helper function recurses down all chart descendants, pruning paths that
are not related to the viewpoint. All tiles that are found are then
passed to a specified tile-drawing funcion.

This is used twice; once for terrain (including roads and rivers), and
again for locations, so that locations are drawn on top of roads.