Refactor to put Coordinates on the stack (#18)
There's no need to malloc data like this (especially now that I have
learned about the relative speeds of mathematics versus cache
retrievals, meaning that it's faster to calculate with small amounts of
data over accessing the end of a pointer).

This also simplifies the code, cutting the number of specialist
Coordinate functions in half or so.